projects/image-moderation/pyproject.toml
[project]
name = "image-moderation"
version = "0.1.0"
description = "Image moderation on LLM Vision models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlx-vlm>=0.1.6",
    "mlx>=0.21.1",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
]
